Basic Teleporting Movement
When the player presses the trigger while aiming the ground with "Teleport" Layer, the player activates the teleportation pointer, on second press while the pointer is visible player will teleport to the pointed location. teleportation pointer automatically disappears when not pointing teleport area or not pressing the trigger for a while. Works on both controllers.